+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
 
  * Add `TesterTrait::assertCommandIsSuccessful()` to test command
  * Deprecate `HelperSet::setCommand()` and `getCommand()` without replacement
+ * Add `rgb(r, g, b)` notation support for output colors
 
 5.3
 ---