+++ src/Symfony/Component/DomCrawler/CHANGELOG.md
@@ -6,6 +6,11 @@ CHANGELOG
 
  * Remove `Crawler::parents()` method, use `ancestors()` instead
 
+5.4
+---
+
+ * Add `Crawler::innerText` method.
+
 5.3
 ---